+++ AbracodeFramework/CocoaControls/OMCView.m
@@ -84,8 +84,8 @@
 		NSArray *subViewsArray = [inView subviews];
 		if(subViewsArray != NULL)
 		{
-			int subIndex;
-			int subCount = [subViewsArray count];
+            NSUInteger subIndex;
+            NSUInteger subCount = [subViewsArray count];
 			for(subIndex = 0; subIndex < subCount; subIndex++)
 			{
 				id viewObject = [subViewsArray objectAtIndex:subIndex];
